Moved recycler items to other package.
